conditional rendering  of report page
